--- app/controllers/api/mappings_controller.rb
@@ -1,2 +1,0 @@
-class Api::MappingsController < API::RestfulController
-end
