--- app/controllers/api/maps_controller.rb
@@ -1,2 +1,0 @@
-class Api::MapsController < API::RestfulController
-end
