--- app/controllers/api/synapses_controller.rb
@@ -1,2 +1,0 @@
-class Api::SynapsesController < API::RestfulController
-end
