--- app/controllers/api/topics_controller.rb
@@ -1,2 +1,0 @@
-class Api::TopicsController < API::RestfulController
-end
