+++ app/controllers/api/v1/mappings_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class MappingsController < RestfulController
+    end
+  end
+end
