+++ app/controllers/api/v1/maps_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class MapsController < RestfulController
+    end
+  end
+end
