+++ app/controllers/api/v1/synapses_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class SynapsesController < RestfulController
+    end
+  end
+end
