+++ app/controllers/api/v1/topics_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class TopicsController < RestfulController
+    end
+  end
+end
