+++ app/serializers/api/v1/metacode_serializer.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    class MetacodeSerializer < ActiveModel::Serializer
+      attributes :id,
+        :name,
+        :manual_icon,
+        :color,
+        :aws_icon
+    end
+  end
+end
