+++ app/serializers/api/v1/synapse_serializer.rb
@@ -1,0 +1,17 @@
+module Api
+  module V1
+    class SynapseSerializer < ActiveModel::Serializer
+      attributes :id,
+        :desc,
+        :category,
+        :weight,
+        :permission,
+        :created_at,
+        :updated_at
+
+      has_one :topic1, root: :topics, serializer: TopicSerializer
+      has_one :topic2, root: :topics, serializer: TopicSerializer
+      has_one :user, serializer: UserSerializer
+    end
+  end
+end
