+++ app/serializers/api/v1/token_serializer.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    class TokenSerializer < ActiveModel::Serializer
+      attributes :id,
+        :token,
+        :description,
+        :created_at,
+        :updated_at
+    end
+  end
+end
