+++ app/serializers/api/v1/topic_serializer.rb
@@ -1,0 +1,16 @@
+module Api
+  module V1
+    class TopicSerializer < ActiveModel::Serializer
+      attributes :id,
+        :name,
+        :desc,
+        :link,
+        :permission,
+        :created_at,
+        :updated_at
+
+      has_one :user, serializer: UserSerializer
+      has_one :metacode, serializer: MetacodeSerializer
+    end
+  end
+end
