+++ app/serializers/api/v1/webhook_serializer.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1
+    class WebhookSerializer < ActiveModel::Serializer
+      attributes :text, :username, :icon_url # , :attachments
+    end
+  end
+end
