--- app/serializers/metacode_serializer.rb
@@ -1,7 +1,0 @@
-class MetacodeSerializer < ActiveModel::Serializer
-  attributes :id,
-             :name,
-             :manual_icon,
-             :color,
-             :aws_icon
-end
