--- app/serializers/token_serializer.rb
@@ -1,7 +1,0 @@
-class TokenSerializer < ActiveModel::Serializer
-  attributes :id,
-             :token,
-             :description,
-             :created_at,
-             :updated_at
-end
