--- app/serializers/webhook_serializer.rb
@@ -1,3 +1,0 @@
-class WebhookSerializer < ActiveModel::Serializer
-  attributes :text, :username, :icon_url # , :attachments
-end
